+++ test_unstructured_ingest/test_interfaces.py
@@ -15,7 +15,7 @@ test_files = [
 
 
 @pytest.mark.parametrize("filename", test_files)
-def test_process_file_include_filename(filename: str):
+def test_process_file_metadata_include_filename(filename: str):
     ingest_doc = GitIngestDoc(
         path=filename,
         config=SimpleGitConfig(
@@ -26,12 +26,11 @@ def test_process_file_include_filename(filename: str):
     isd_elems = ingest_doc.process_file()
 
     for elem in isd_elems:
-        for k in elem["metadata"]:
-            assert k == "filename"
+        assert set(elem["metadata"].keys()) == {"filename"}
 
 
 @pytest.mark.parametrize("filename", test_files)
-def test_process_file_include_filename_pagenum(filename: str):
+def test_process_file_metadata_include_filename_pagenum(filename: str):
     ingest_doc = GitIngestDoc(
         path=filename,
         config=SimpleGitConfig(
@@ -42,12 +41,11 @@ def test_process_file_include_filename_pagenum(filename: str):
     isd_elems = ingest_doc.process_file()
 
     for elem in isd_elems:
-        for k in elem["metadata"]:
-            assert k in ["filename", "page_number"]
+        assert set(elem["metadata"].keys()) == {"filename", "page_number"}
 
 
 @pytest.mark.parametrize("filename", test_files)
-def test_process_file_exclude_filename(filename: str):
+def test_process_file_metadata_exclude_filename(filename: str):
     ingest_doc = GitIngestDoc(
         path=filename,
         config=SimpleGitConfig(
@@ -58,12 +56,11 @@ def test_process_file_exclude_filename(filename: str):
     isd_elems = ingest_doc.process_file()
 
     for elem in isd_elems:
-        for k in elem["metadata"]:
-            assert k != "filename"
+        assert "filename" not in elem["metadata"].keys()
 
 
 @pytest.mark.parametrize("filename", test_files)
-def test_process_file_exclude_filename_pagenum(filename: str):
+def test_process_file_metadata_exclude_filename_pagenum(filename: str):
     ingest_doc = GitIngestDoc(
         path=filename,
         config=SimpleGitConfig(
@@ -74,5 +71,66 @@ def test_process_file_exclude_filename_pagenum(filename: str):
     isd_elems = ingest_doc.process_file()
 
     for elem in isd_elems:
-        for k in elem["metadata"]:
-            assert k not in ["filename", "page_number"]
+        assert "filename" not in elem["metadata"].keys()
+        assert "page_number" not in elem["metadata"].keys()
+
+
+@pytest.mark.parametrize("filename", test_files)
+def test_process_file_fields_include_default(filename: str):
+    ingest_doc = GitIngestDoc(
+        path=filename,
+        config=SimpleGitConfig(
+          download_dir=EXAMPLE_DOCS_DIRECTORY,
+        ),
+    )
+    isd_elems = ingest_doc.process_file()
+
+    for elem in isd_elems:
+        assert {"element_id", "text", "type", "metadata"} == set(elem.keys())
+
+
+@pytest.mark.parametrize("filename", test_files)
+def test_process_file_fields_include_elementid(filename: str):
+    ingest_doc = GitIngestDoc(
+        path=filename,
+        config=SimpleGitConfig(
+          download_dir=EXAMPLE_DOCS_DIRECTORY,
+          fields_include="element_id",
+        ),
+    )
+    isd_elems = ingest_doc.process_file()
+
+    for elem in isd_elems:
+        assert {"element_id"} == set(elem.keys())
+
+
+@pytest.mark.parametrize("filename", test_files)
+def test_process_file_flatten_metadata_filename(filename: str):
+    ingest_doc = GitIngestDoc(
+        path=filename,
+        config=SimpleGitConfig(
+          download_dir=EXAMPLE_DOCS_DIRECTORY,
+          metadata_include="filename",
+          flatten_metadata=True,
+        ),
+    )
+    isd_elems = ingest_doc.process_file()
+    
+    for elem in isd_elems:
+        assert {"element_id", "text", "type", "filename"} == set(elem.keys())
+
+
+@pytest.mark.parametrize("filename", test_files)
+def test_process_file_flatten_metadata_filename_pagenum(filename: str):
+    ingest_doc = GitIngestDoc(
+        path=filename,
+        config=SimpleGitConfig(
+          download_dir=EXAMPLE_DOCS_DIRECTORY,
+          metadata_include="filename,page_number",
+          flatten_metadata=True,
+        ),
+    )
+    isd_elems = ingest_doc.process_file()
+    
+    for elem in isd_elems:
+        assert {"element_id", "text", "type", "filename", "page_number"} == set(elem.keys())