+++ unstructured/ingest/connector/git.py
@@ -28,6 +28,8 @@ class SimpleGitConfig(BaseConnectorConfig):
     re_download: bool = False
     metadata_include: Optional[str] = None
     metadata_exclude: Optional[str] = None
+    fields_include: str = "element_id,text,type,metadata"
+    flatten_metadata: bool = False
 
     repo_path: str = field(init=False, repr=False)