+++ unstructured/ingest/connector/google_drive.py
@@ -79,6 +79,8 @@ class SimpleGoogleDriveConfig(BaseConnectorConfig):
     preserve_downloads: bool = False
     metadata_include: Optional[str] = None
     metadata_exclude: Optional[str] = None
+    fields_include: str = "element_id,text,type,metadata"
+    flatten_metadata: bool = False
 
     recursive: bool = False
 
@@ -86,9 +88,8 @@ class SimpleGoogleDriveConfig(BaseConnectorConfig):
         if self.extension and self.extension not in EXT_TO_FILETYPE.keys():
             raise ValueError(
                 f"Extension not supported. "
-                f"Value MUST be one of {', '.join(EXT_TO_FILETYPE.keys())}.",
+                f"Value MUST be one of {', '.join([k for k in EXT_TO_FILETYPE if k is not None])}.",
             )
-
         self.service = create_service_account_object(self.service_account_key, self.drive_id)