+++ unstructured/ingest/connector/reddit.py
@@ -33,6 +33,8 @@ class SimpleRedditConfig(BaseConnectorConfig):
     re_download: bool = False
     metadata_include: Optional[str] = None
     metadata_exclude: Optional[str] = None
+    fields_include: str = "element_id,text,type,metadata"
+    flatten_metadata: bool = False
 
     def __post_init__(self):
         if self.num_posts <= 0: