+++ unstructured/ingest/interfaces.py
@@ -50,6 +50,8 @@ class BaseConnectorConfig(ABC):
     re_download: bool = False
     metadata_include: Optional[str] = None
     metadata_exclude: Optional[str] = None
+    fields_include: str = "element_id,text,type,metadata"
+    flatten_metadata: bool = False
 
 
 class BaseIngestDoc(ABC):
@@ -117,7 +119,14 @@ class BaseIngestDoc(ABC):
                     if k not in in_list:
                         elem["metadata"].pop(k, None)  # type: ignore[attr-defined]
 
-            elem.pop("coordinates")  # type: ignore[attr-defined]
+            in_list = self.config.fields_include.split(",")
+            elem = {k: v for k, v in elem.items() if k in in_list}
+
+            if self.config.flatten_metadata:
+                for k, v in elem["metadata"].items():  # type: ignore[attr-defined]
+                    elem[k] = v
+                elem.pop("metadata")  # type: ignore[attr-defined]
+
             self.isd_elems_no_filename.append(elem)
 
         return self.isd_elems_no_filename