+++ unstructured/partition/md.py
@@ -25,6 +25,8 @@ def partition_md(
     parser: VALID_PARSERS = None,
 ) -> List[Element]:
     # Verify that only one of the arguments was provided
+    if text is None:
+        text = ""
     exactly_one(filename=filename, file=file, text=text, url=url)
 
     if filename is not None:
@@ -44,9 +46,6 @@ def partition_md(
             raise ValueError(f"Expected content type text/markdown. Got {content_type}.")
 
         text = response.text
-
-    elif text is None:
-        raise ValueError("Exactly one of filename, file, url or text must be specified.")
 
     html = markdown.markdown(text)